Решение для 8_7_6 kyu
